Refactoring of assert function
