+++ tangelo/linq/tests/test_ibm_connection.py
@@ -9,6 +9,7 @@ import time
 
 import numpy as np
 
+from tangelo.helpers.utils import assert_freq_dict_almost_equal
 from tangelo.linq import Gate, Circuit, Simulator
 from tangelo.linq.qpu_connection import IBMConnection
 from tangelo.toolboxes.operators import QubitOperator
@@ -24,18 +25,7 @@ sim = Simulator()
 ref_estimator = sim.get_expectation_value(op, circ2)
 
 
-def assert_freq_dict_almost_equal(d1, d2, atol):
-    """ Utility function to check whether two frequency dictionaries are almost equal, for arbitrary tolerance """
-    if d1.keys() != d2.keys():
-        raise AssertionError("Dictionary keys differ. Frequency dictionaries are not almost equal.\n"
-                             f"d1 keys: {d1.keys()} \nd2 keys: {d2.keys()}")
-    else:
-        for k in d1.keys():
-            if abs(d1[k] - d2[k]) > atol:
-                raise AssertionError(f"Frequency {k}, difference above tolerance {atol}: {d1[k]} != {d2[k]}")
-
-
-os.environ['IBM_TOKEN'] = 'insert valid token'
+os.environ['IBM_TOKEN'] = 'insert valid token here'
 @unittest.skip("We do not want to store login information for automated testing. Manual testing only.")
 class TestIBMConnection(unittest.TestCase):
 
@@ -49,6 +39,7 @@ class TestIBMConnection(unittest.TestCase):
         tmp = os.getenv("IBM_TOKEN", '')
         os.environ['IBM_TOKEN'] = 'invalid_apikey'
         self.assertRaises(RuntimeError, IBMConnection)
+        self.assertRaises(RuntimeError, IBMConnection, 'invalid')
         os.environ['IBM_TOKEN'] = ''
         self.assertRaises(RuntimeError, IBMConnection)
         os.environ['IBM_TOKEN'] = tmp
